Breaks down fetching into func modules
