+++ src/utils/fetchPush.js
@@ -1,6 +1,10 @@
-export const fetchPush = async (urlWithCat) => {
+export const fetchPush = async (category) => {
+  const url = `https://en.wikipedia.org/w/api.php?action=query&list=categorymembers&cmtitle=Category:${category.replace(
+    /[" "]/g,
+    "_"
+  )}&cmprop=title|type&format=json&cmlimit=500&cmtype=page|subcat&origin=*`; // cannot lower case URLs, Wikipedia categories are case sensitive
   const fetchedData = { articles: [], subCats: [] };
-  const urlFetch = await fetch(urlWithCat);
+  const urlFetch = await fetch(url);
   const {
     query: { categorymembers },
   } = await urlFetch.json();
